Add build that preserves stringification.
For use with fast-async, it is important that both $asyncbind
and $asyncspawn can be stringified and have no dependencies.

To do so, we require the contents of their dependencies and
inline them.

Both unminified and minified source are written. By default,
the minified source is used.

This reduces runtime size slightly (due to the removal of
processIncludes()) and no longer uses eval, making it compatible
with a strict Content-Security-Policy.

Fixes #3